+++ projects/lib/src/lib/forms/form-blocks/form-block-field/form-block-field-floating-action/form-block-field-floating-action.component.ts
@@ -1,10 +1,14 @@
+import { CommonModule } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
 
 @Component({
     selector: 'ce-form-block-field-floating-action',
     templateUrl: './form-block-field-floating-action.component.html',
     styleUrls: ['./form-block-field-floating-action.component.scss'],
-    standalone: false
+    imports: [
+      CommonModule,
+    ],
+    standalone: true,
 })
 export class FormBlockFieldFloatingActionComponent implements OnInit {