+++ projects/lib/src/lib/forms/form-blocks/form-block-field/form-block-field-infos/form-block-field-infos.component.ts
@@ -1,0 +1,14 @@
+import { CommonModule } from '@angular/common';
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'ce-form-block-field-infos',
+  imports: [
+    CommonModule,
+  ],
+  templateUrl: './form-block-field-infos.component.html',
+  styleUrl: './form-block-field-infos.component.scss'
+})
+export class FormBlockFieldInfosComponent {
+
+}