+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [19.1.3]
+ - add api documentation
+ - update doc configuration bootstrap
+ - support for form block array query params
+ - add filters label component
+ - refacto : more standalone modules
+ - add filters infos for form array block
+
 ## [19.1.2]
  - Fix: add readonly support for checkboxes